everpal: perfmgr: Add in TOUCH hints to INTERACTION hints
MTK APP TOUCH hints are triggered when an app is being used to
boost the responsiveness of the app.
But MTK APP TOUCH hints are non-existent on AOSP as they rely on MTK
boost framework. On the other hand AOSP provides handy boosting hints
by default with the AIDL interfaces.
These are available in the stock powerhal but are not enabled by
default because of the different approach of the MTK Boosting
framework.
Add these app touch hints to INTERACTION, making the UX even better.

co-authored-by: Vaisakh Murali <mvaisakh@statixos.com>
Signed-off-by: weaponmasterjax <42597871+weaponmasterjax@users.noreply.github.com>